# docs/05-openapi.yaml
openapi: 3.1.0
info:
  title: TransitGuard API
  version: 1.0.0 (Entwurf, implementierungsreif)
  description: >
    ÖPNV-Companion-API. Meldungs-Ressourcen sind no-store. Auth via X-Device-Token;
    Pro-Endpunkte via Bearer-Entitlement-Token. Vollständige Semantik: docs/04-api.md.
servers:
  - url: https://api.example.de
  - url: https://api.example.de/{basePath}
    variables: { basePath: { default: v1, enum: [v1] } }
security:
  - deviceToken: []
...
paths:
...
  /v1/reports:
    post:
      tags: [reports]
      parameters: [ { name: Idempotency-Key, in: header, required: true, schema: { type: string, format: uuid } } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ReportCreate" } } }
      responses:
        "201": { description: erstellt, content: { application/json: { schema: { $ref: "#/components/schemas/ReportFree" } } } }
        "422": { $ref: "#/components/responses/ValidationError" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /v1/reports/{id}/events:
    post:
      tags: [reports]
      parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [type], properties: { type: { type: string, enum: [confirm, contradict] } } } } }
      responses: { "204": { description: registriert }, "429": { $ref: "#/components/responses/RateLimited" } }
  /v1/cities/{slug}/reports:
    get:
      tags: [reports]
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
        - { name: bbox, in: query, description: "minLon,minLat,maxLon,maxLat", schema: { type: string } }
        - { name: active, in: query, schema: { type: boolean, default: true } }
        - { name: since_id, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer, default: 100, maximum: 200 } }
      responses:
        "200": { description: Tier-geformte Liste, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/ReportFree" } } } } }
  /v1/cities/{slug}/events:
    get:
      tags: [reports]
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
        - { name: since_id, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer, default: 200, maximum: 1000 } }
      responses:
        "200": { description: Ereignisstrom ab Kursor, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/ReportEvent" } } } } }
...
  /v1/billing/activate:
    post:
      tags: [billing]
      security: []
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [platform, receipt], properties: { platform: { type: string, enum: [apple, google] }, receipt: { type: string } } } } }
      responses:
        "201": { description: Entitlement ausgestellt, content: { application/json: { schema: { $ref: "#/components/schemas/EntitlementIssued" } } } }
        "402": { description: receipt_invalid }
        "409": { description: receipt_already_used }
  /v1/billing/restore:
    post:
      tags: [billing]
      security: []
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [restore_code], properties: { restore_code: { type: string } } } } }
      responses: { "201": { description: neues Token, content: { application/json: { schema: { $ref: "#/components/schemas/EntitlementIssued" } } } }, "403": { description: restore_invalid } }
  /v1/billing/validate:
    post:
      tags: [billing]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [token], properties: { token: { type: string } } } } }
      responses: { "200": { description: Status, content: { application/json: { schema: { $ref: "#/components/schemas/EntitlementStatus" } } } } }
components:
  securitySchemes:
    deviceToken: { type: apiKey, in: header, name: X-Device-Token }
    entitlementToken: { type: http, scheme: bearer, bearerFormat: "et_…" }
  responses:
    ValidationError: { description: 422, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    RateLimited: { description: 429, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  schemas:
    Error:
      type: object
      properties: { code: { type: string }, message: { type: string }, details: { type: object } }
...
    ReportCreate:
      type: object
      required: [city_slug, anchor_type, report_type, vehicle_kind, station]
      properties:
        city_slug: { type: string }
        anchor_type: { type: string, enum: [station, trip] }
        report_type: { type: string, enum: [in_vehicle, on_platform, at_stop] }
        vehicle_kind: { type: string, enum: [rail, bus] }
        station: { type: object, required: [stop_id], properties: { stop_id: { type: string }, geo: { $ref: "#/components/schemas/LatLon" } } }
        trip: { type: object, properties: { trip_id: { type: string }, start_date: { type: string }, route_id: { type: string }, direction_id: { type: integer }, headsign: { type: string } } }
        inspector_count: { type: integer, minimum: 1, maximum: 8 }
        client:
          type: object
          required: [ticket_confirmed]
          properties:
            ticket_confirmed: { type: boolean, description: 'Pflicht: Bestätigung gültiger Fahrschein (Ticket-First-Gate, docs/18)' }
            app_version: { type: string }
            platform: { type: string, enum: [ios, android, web] }
    LatLon: { type: object, required: [lat, lon], properties: { lat: { type: number }, lon: { type: number } } }
    ReportFree:
      type: object
      description: Free-Form — Pro-Felder fehlen physisch (Kontrakt, docs/04 §3)
      properties:
        id: { type: string, format: uuid }
        city_slug: { type: string }
        anchor_type: { type: string, enum: [station, trip, derived_terminus] }
        station_name: { type: string }
        route_id: { type: string }
        headsign: { type: string }
        report_type: { type: string }
        created_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        status: { type: string }
        origin: { type: string, enum: [user, derived] }
        trip_id: { type: string }
        trip_start_date: { type: string }
...
    ReportEvent:
      type: object
      properties:
        event_id: { type: integer }
        report_id: { type: string, format: uuid }
        type: { type: string, enum: [created, confirmed, contradicted, ttl_extended, degraded, expired, cancelled] }
        created_at: { type: string, format: date-time }
        payload: { type: object }
...
    EntitlementIssued:
      type: object
      properties: { access_token: { type: string }, restore_code: { type: string }, tier: { type: string }, valid_until: { type: [string, "null"], format: date-time } }
    EntitlementStatus:
      type: object
      properties: { tier: { type: string, enum: [free, pro] }, valid_until: { type: [string, "null"], format: date-time }, status: { type: string } }